float t, color max_to_one

+++ Objeto.h
@@ -15,7 +15,7 @@ float kd, ks;
 Objeto(vec3f col):color{col}{}
 
 //virtual bool intersectar(Rayo ray, int &t, vec3f &c)=0;
-virtual bool intersectar(Rayo ray, int &t, vec3f &c, vec3f &normal)=0;
+virtual bool intersectar(Rayo ray, float &t, vec3f &c, vec3f &normal)=0;
 };
 
 class Esfera : public Objeto {
@@ -24,15 +24,15 @@ public:
     float radio;
 
     Esfera(vec3f cen, float r, vec3f col): centro{cen}, radio{r}, Objeto(col) {}
-    bool intersectar(Rayo ray, int &t, vec3f &col, vec3f &normal) {
+    bool intersectar(Rayo ray, float &t, vec3f &col, vec3f &normal) {
         auto _a = ray.dir.productoPunto(ray.dir);
         auto _b = 2*ray.dir.productoPunto(ray.ori-centro);
         auto _c = (ray.ori-centro).productoPunto(ray.ori-centro)-radio*radio;
         auto D = _b*_b-4*_a*_c;
         if(D <= 0) {return false;}
-        int t1 = (-_b + sqrt(D))/2*_a;
-        int t2 = (-_b - sqrt(D))/2*_a;
-        t = std::min(t1,t2);
+        float t1 = (-_b + sqrt(D))/2*_a;
+        float t2 = (-_b - sqrt(D))/2*_a;
+        t = std::min(t1, t2);
         if(t <= 0) {return false;}
         col = color;
         vec3f pi = ray.punto_interseccion(t);
